requirements.txt
numpy>=2.0
faiss-cpu==1.12.0
fastapi==0.115.4
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
# numpy==1.26.4
# faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.3.1
python-docx==1.1.2
markdown-it-py==3.0.0

# requirements.in  (allow patch updates)
# pip install pip-tools
# pip-compile requirements.in -o requirements.txt     # creates a fully pinned lock
# pip-sync requirements.txt                           # installs exactly the lock
# # Later upgrades:
# pip-compile --upgrade requirements.in -o requirements.txt
